+++ src/folding_range.rs
@@ -57,7 +57,7 @@ pub struct FoldingRangeCapability {
     /// The maximum number of folding ranges that the client prefers to receive per document. The value serves as a
     /// hint, servers are free to follow the limit.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub range_limit: Option<u64>,
+    pub range_limit: Option<u32>,
     /// If set, the client signals that it only supports folding complete lines. If set, client will
     /// ignore specified `startCharacter` and `endCharacter` properties in a FoldingRange.
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -81,18 +81,18 @@ pub enum FoldingRangeKind {
 #[serde(rename_all = "camelCase")]
 pub struct FoldingRange {
     /// The zero-based line number from where the folded range starts.
-    pub start_line: u64,
+    pub start_line: u32,
 
     /// The zero-based character offset from where the folded range starts. If not defined, defaults to the length of the start line.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub start_character: Option<u64>,
+    pub start_character: Option<u32>,
 
     /// The zero-based line number where the folded range ends.
-    pub end_line: u64,
+    pub end_line: u32,
 
     /// The zero-based character offset before the folded range ends. If not defined, defaults to the length of the end line.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub end_character: Option<u64>,
+    pub end_character: Option<u32>,
 
     /// Describes the kind of the folding range such as `comment' or 'region'. The kind
     /// is used to categorize folding ranges and used by commands like 'Fold all comments'. See
